regsiter page errors got managed
